Add a `useLocalUsers` hook

+++ src/components/scenes/PasswordLoginScene.tsx
@@ -17,8 +17,8 @@ import { launchPasswordRecovery, login } from '../../actions/LoginAction'
 import { maybeRouteComplete } from '../../actions/LoginInitActions'
 import s from '../../common/locales/strings'
 import { useImports } from '../../hooks/useImports'
+import { LoginUserInfo, useLocalUsers } from '../../hooks/useLocalUsers'
 import { BiometryType } from '../../keychain'
-import { LoginUserInfo } from '../../reducers/PreviousUsersReducer'
 import { Branding } from '../../types/Branding'
 import { useDispatch, useSelector } from '../../types/ReduxTypes'
 import { SceneProps } from '../../types/routerTypes'
@@ -114,14 +114,16 @@ class PasswordLoginSceneComponent extends React.Component<Props, State> {
 
   handleDelete = (userInfo: LoginUserInfo) => {
     const { context } = this.props
-    const { username } = userInfo
 
     Keyboard.dismiss()
     Airship.show(bridge => (
       <ButtonsModal
         bridge={bridge}
         title={s.strings.forget_account}
-        message={sprintf(s.strings.forget_username_account, username)}
+        message={sprintf(
+          s.strings.forget_username_account,
+          userInfo.username ?? s.strings.username
+        )}
         buttons={{
           ok: { label: s.strings.forget },
           cancel: { label: s.strings.cancel, type: 'secondary' }
@@ -130,7 +132,13 @@ class PasswordLoginSceneComponent extends React.Component<Props, State> {
     ))
       .then(async button => {
         if (button !== 'ok') return
-        return await context.deleteLocalAccount(username)
+        if (context.forgetAccount != null) {
+          await context.forgetAccount(userInfo.loginId)
+        } else {
+          const { username } = userInfo
+          if (username == null) throw new Error('No username')
+          await context.deleteLocalAccount(username)
+        }
       })
       .catch(showError)
   }
@@ -310,6 +318,7 @@ class PasswordLoginSceneComponent extends React.Component<Props, State> {
 
   handleSelectUser = (userInfo: LoginUserInfo) => {
     const { username } = userInfo
+    if (username == null) return // These don't exist in the list
     this.handleChangeUsername(username)
     this.setState({
       usernameList: false
@@ -320,7 +329,8 @@ class PasswordLoginSceneComponent extends React.Component<Props, State> {
     )
     if (
       details != null &&
-      (details.pinEnabled || (details.touchEnabled && this.props.touch))
+      (details.pinLoginEnabled ||
+        (details.touchLoginEnabled && this.props.touch))
     ) {
       this.props.gotoPinLoginPage()
       return
@@ -414,7 +424,7 @@ export function PasswordLoginScene(props: OwnProps) {
   const dispatch = useDispatch()
   const theme = useTheme()
 
-  const localUsers = useSelector(state => state.previousUsers.userList)
+  const localUsers = useLocalUsers()
   const touch = useSelector(state => state.touch.type)
   const username = useSelector(state => state.login.username)
